+++ src/levels/Level.java
@@ -6,8 +6,12 @@ public class Level {
     public Level(int[][] levelData) {
         this.levelData = levelData;
     }
-
+    
     public int getSpriteIndex(int x, int y) {
         return levelData[y][x];
+    }
+    
+    public int[][] getLevelData() {
+        return levelData;
     }
 }